+++ tools/aa/include/itest_observer.h
@@ -31,7 +31,7 @@ public:
      * @param msg, Indicates the status information.
      * @param resultCode, Indicates the result code.
      */
-    virtual void TestStatus(const std::string &msg, const int &resultCode) = 0;
+    virtual void TestStatus(const std::string &msg, const int64_t &resultCode) = 0;
 
     /**
      * Outputs information and result code that the test has finished.
@@ -39,7 +39,7 @@ public:
      * @param msg, Indicates the status information.
      * @param resultCode, Indicates the result code.
      */
-    virtual void TestFinished(const std::string &msg, const int &resultCode) = 0;
+    virtual void TestFinished(const std::string &msg, const int64_t &resultCode) = 0;
 
     /**
      * Executes the specified shell command.
